MultiAgentSysAdmin

MultiAgentSysAdmin.AbstractSysAdmin

MultiAgentSysAdmin.MachineState

POMDPs.gen

MultiAgentSysAdmin.AbstractSysAdmin — Type

p_fail_base:

p_fail_bonus:

p_dead_base:

p_dead_bonus:

p_load: probability of getting a job when idle.

p_doneG: probability of completing a job when good.

p_doneF: probability of completing a job when faulty.

p_fail_bonus and p_dead_bonus are additional bonuses counted when all neighbors are faulty. Counted per agent. If a machine with 2 neighbors has a single faulty neighbor, it will get an additional failing probability of p_fail_bonus/2. If the same machine has one faulty neighbor and one dead neighbor, it will get a penalty of p_fail_bonus/2 + p_dead_bonus/2.

MultiAgentSysAdmin.MachineState — Type

status: {good, faulty, dead}

load: {idle, loaded, success}

POMDPs.gen — Method

Basically, the only way we can get reward is by:

Starting from the Load state (since it's the only one that can complete)

Doing action 0;

And ending up in the Done state.

dead machine increases the probability that its neighbors become faulty and die system receives a reward of 1 if a process terminates successfully status is faulty, processes take longer to terminate If the machine dies, the process is lost.